refactor(api): remove event context
