+++ webclient/package.json
@@ -16,7 +16,7 @@
     "react-router-dom": "^5.2.0"
   },
   "devDependencies": {
-    "@types/node": "^13.13.4",
+    "@types/node": "^14.11.2",
     "@types/react": "^16.4.18",
     "@types/react-dom": "^16.0.9",
     "@types/react-router-dom": "^5.1.5",
